requirements.txt
streamlit
firebase_admin
mysql-connector-python
numpy
pandas